refactor: normalize storybook folder structure
