+++ frontend/src/stories/Navigation/Sidebar.stories.tsx
