+++ frontend/src/stories/Primitives/Avatar/Avatar.stories.tsx
