+++ frontend/src/stories/Primitives/Button.stories.tsx
@@ -12,7 +12,7 @@ import {
   ButtonSizeType,
   ButtonVariant,
   ButtonSize,
-} from './types/PrimitiveTypes';
+} from '../types/PrimitiveTypes';
 
 const VARIANT_OPTIONS: ButtonVariantType[] = [...ButtonVariant];
 const SIZE_OPTIONS: ButtonSizeType[] = [...ButtonSize];
