+++ frontend/src/stories/Primitives/Loading/Dots.stories.tsx
@@ -1,31 +1,31 @@
 import React from 'react';
 import { ComponentStory } from '@storybook/react';
 
-import Spinner from '@/components/Primitives/Loading/Spinner';
+import Dots from '@/components/Primitives/Loading/Dots';
 import dedent from 'ts-dedent';
-import { SpinnerColorType, SpinnerSizeType } from '../types/PrimitiveTypes';
+import { DotsColorType, DotsSizeType } from '../../types/PrimitiveTypes';
 
-const COLOR_OPTIONS: SpinnerColorType[] = ['light', 'dark'];
-const SIZE_OPTIONS: SpinnerSizeType[] = [50, 80, 100, 150, 200];
+const COLOR_OPTIONS: DotsColorType[] = ['primary800', 'primary200', 'white'];
+const SIZE_OPTIONS: DotsSizeType[] = [8, 4, 10, 15, 50, 80, 100];
 
 export default {
-  title: 'Primitives/Loading/Spinner',
-  component: Spinner,
+  title: 'Primitives/Loading/Dots',
+  component: Dots,
   parameters: {
     docs: {
       description: {
         component: dedent`
-        Spinner element, it spins forever.
+        Dots element, it's used for Loading.
 
         **File Path:**
-        \`@/components/Primitives/Loading/Spinner.tsx\`
+        \`@/components/Primitives/Loading/Dots/index.tsx\` and \`@/components/Primitives/Loading/Dots/styles.tsx\`
         `,
       },
     },
   },
   args: {
-    size: 80,
-    color: 'dark',
+    size: 15,
+    color: 'primary800',
   },
   argTypes: {
     color: {
@@ -34,7 +34,7 @@ export default {
       description: 'The component color.',
       table: {
         type: { summary: COLOR_OPTIONS.join('|') },
-        defaultValue: { summary: 'dark' },
+        defaultValue: { summary: 'primary800' },
       },
     },
     size: {
@@ -43,13 +43,13 @@ export default {
       description: 'The component size.',
       table: {
         type: { summary: SIZE_OPTIONS.join('|') },
-        defaultValue: { summary: 80 },
+        defaultValue: { summary: 15 },
       },
     },
   },
 };
 
-const Template: ComponentStory<typeof Spinner> = ({ ...args }) => <Spinner {...args} />;
+const Template: ComponentStory<typeof Dots> = ({ ...args }) => <Dots {...args} />;
 
 export const Default = Template.bind({});
 Default.storyName = 'Basic Usage';
