+++ frontend/src/stories/Primitives/Svg.stories.tsx
@@ -5,7 +5,7 @@ import dedent from 'ts-dedent';
 
 import Svg from '@/components/Primitives/Svg';
 import Flex from '@/components/Primitives/Flex';
-import { SvgSizeType } from './types/PrimitiveTypes';
+import { SvgSizeType } from '../types/PrimitiveTypes';
 
 const SIZE_OPTIONS: SvgSizeType[] = [12, 16, 18, 20, 24, 32];
